Move solve rewrites under tensor/rewriting/linalg
The scan-tracking solve rewrites previously lived in
pytensor/tensor/_linalg/solve/rewriting.py and were registered via a
side-effect import chain from tensor/_linalg/__init__.py. That chain
was broken in the linalg reorg (the _linalg package init was emptied),
and registration only kept working incidentally because
tensor/linalg.py happens to import from _linalg.solve.

Move them into tensor/rewriting/linalg/solve.py alongside the other
solve rewrites. The imports also needed updating: the v3 file imported
via tensor.slinalg, which is now the deprecation shim and would emit
DeprecationWarnings on every load.

Also drop the speculative __all__ and inv_to_solve re-export from
tensor/rewriting/linalg/__init__.py; the one test that imported
inv_to_solve from there now imports it from .inverse directly.
