--- pytensor/tensor/_linalg/solve/__init__.py
@@ -1,2 +1,0 @@
-# Register rewrites in the database
-import pytensor.tensor._linalg.solve.rewriting
