+++ pytensor/tensor/rewriting/linalg/__init__.py
@@ -3,25 +3,9 @@
 Importing this package registers all linalg rewrites with the optimizer.
 """
 
-from pytensor.tensor.rewriting.linalg import (
-    decomposition,
-    inverse,
-    products,
-    solve,
-    summary,
-    utils,
-)
-
-# Re-export for backwards compatibility
-from pytensor.tensor.rewriting.linalg.inverse import inv_to_solve
-
-
-__all__ = [
-    "decomposition",
-    "inv_to_solve",
-    "inverse",
-    "products",
-    "solve",
-    "summary",
-    "utils",
-]
+import pytensor.tensor.rewriting.linalg.decomposition
+import pytensor.tensor.rewriting.linalg.inverse
+import pytensor.tensor.rewriting.linalg.products
+import pytensor.tensor.rewriting.linalg.solve
+import pytensor.tensor.rewriting.linalg.summary
+import pytensor.tensor.rewriting.linalg.utils
